requirements.txt
# Auto-generated requirements.txt
bs4==0.0.1
python-dotenv==0.21.1
requests==2.28.1
selenium==4.8.0
webdriver_manager==4.0.2
